refactor: pulling some small composables out of history events
